refactor: split controller & utils
