+++ constants.js
@@ -1,4 +1,4 @@
 export const cmdOverloaded =
   "This command does not require any additional arguments, so extra arguments are ignored\n";
 export const lackOfArgs =
-  "This command require additional arguments, but they are absent\n";
+  "Invalid input. This command require additional arguments, but they are absent\n";
